+++ platform/cc2530eb/drivers/hmc5883.h
@@ -4,9 +4,9 @@
 #define HMC5883_SLAVE_ADDR 0x3C
 
 typedef struct hmc5883_data{
-    int x;
-    int y;
-    int z;
+    int16_t x;
+    int16_t y;
+    int16_t z;
 }HMC5883_DATA;
